CUSTOM-204 Adds an option to configure the log verbosity
https://perconadev.atlassian.net/browse/CUSTOM-204

By default, proxysql console log contains all types of messages
including info level messages for diagnostic purposes. But, sometimes
this hinders the readability of the logfile.

So, a new option, "log_verbosity_level" is introduced which
helps in configuring the extent of verbosity in the logs.

Depending on the value set for the option from the configuration file
globally,

```
   .
   .
   datadir="/home/varun/proxysql/proxysql-sysown/data"
   log_verbosity_level=3
   .
   .
   admin_variables=
   {
   .
   .
```

The behaviour of proxysql logging is as follows

log_verbosity_level = 1 | ERROR
log_verbosity_level = 2 | ERROR, WARNING
log_verbosity_level = 3 | ERROR, WARNING, INFORMATION

By default, the value of the option is set to 3, meaning Errors,
warnings and info messages are logged if not excpliclty configured.

Note:-
This patch is adapted from the PR #4477.
https://github.com/sysown/proxysql/pull/4477/files