CUSTOM-204 Avoid printing SQL queries into ProxySQL log file
https://perconadev.atlassian.net/browse/CUSTOM-204

In some cases, it is required by the database customers to limit their
system administrator with access to ProxySQL log files from viewing
the SQL statements since they may contain sensitive data.

So, a new MySQL variable, "log_mysql_statements" is introduced which
controls whether the SQL statement would be printed to the proxysql log
file or not.

``` .
    .
    mysql_variables=
    {
	threads=4
	max_connections=2048
	log_mysql_statements=false
    .
    .
    .
```

By default, it will be TRUE, meaning the SQL statements
would be printed regardless of if it is a info message, warning, or an
error. If set to false, the SQL statements will be redacted with an
alternate placeholder text. Also, since it is a MySQL variable it is
dynamically modifyable.